requirements.txt
PyGithub~=2.6.1
tqdm~=4.67.1
colorama~=0.4.6